move request factories to request.go
