+++ sync/worker.go
@@ -102,56 +102,3 @@ func NewNeighborhoodWorker(s *server.MessageServer,
 	}
 
 }
-
-func LayerIdsReqFactory(lyr types.LayerID) RequestFactory {
-	return func(s *server.MessageServer, peer p2p.Peer) (chan interface{}, error) {
-		ch, foo := layerBlockIDsRequest()
-		if err := s.SendRequest(LAYER_IDS, lyr.ToBytes(), peer, foo); err != nil {
-			log.Error("could not get layer ", lyr, " hash from peer ", peer)
-			return nil, errors.New("error ")
-		}
-		return ch, nil
-	}
-}
-
-func HashReqFactory(lyr types.LayerID) RequestFactory {
-	return func(s *server.MessageServer, peer p2p.Peer) (chan interface{}, error) {
-		log.Info("send layer hash request Peer: %v layer: %v", peer, lyr)
-		ch, foo := layerHashRequest(peer)
-		if err := s.SendRequest(LAYER_HASH, lyr.ToBytes(), peer, foo); err != nil {
-			s.Error("could not get layer ", lyr, " hash from peer ", peer)
-			return nil, errors.New("error ")
-		}
-
-		return ch, nil
-	}
-
-}
-
-//todo handle blocks in retry queue
-func BlocReqFactory(id types.BlockID) RequestFactory {
-	return func(s *server.MessageServer, peer p2p.Peer) (chan interface{}, error) {
-		log.Info("send block request Peer: %v layer: %v", peer, id)
-		ch, foo := miniBlockRequest()
-		if err := s.SendRequest(BLOCK, id.ToBytes(), peer, foo); err != nil {
-			s.Error("could not get block ", id, " hash from peer ", peer)
-			return nil, errors.New("error ")
-		}
-
-		return ch, nil
-	}
-}
-
-//todo batch requests
-func TxReqFactory(id types.TransactionId) RequestFactory {
-	return func(s *server.MessageServer, peer p2p.Peer) (chan interface{}, error) {
-		log.Info("send block request Peer: %v layer: %v", peer, id)
-		ch, foo := txRequest()
-		if err := s.SendRequest(TX, id[:], peer, foo); err != nil {
-			s.Error("could not get transaction ", id, "  from peer ", peer)
-			return nil, errors.New("error ")
-		}
-
-		return ch, nil
-	}
-}
